refactor: util 폴더를 class화 하여 리팩토링
