--- utils/contactData.tsx
@@ -1,6 +1,0 @@
-export default [
-  { title: 'Leader', value: '김성민', hasLink: false },
-  { title: 'Github', value: 'https://github.com/Collection50', hasLink: true },
-  { title: 'Phone', value: '010 - 1234 - 5678', hasLink: false },
-  { title: 'Open Kakao', value: 'https://open.kakao.com/o/sObD2U5e', hasLink: true },
-];
