--- utils/mobilePromotionDatas.tsx
@@ -1,28 +1,0 @@
-export default [
-  {
-    title: 'Project',
-    height: 176,
-    benefits: [
-      '웹페이지, 모바일 앱을 만들어 배포하며\n실무 경험을 쌓아요.',
-      '6주간의 프로젝트를 통해 분야에 대한\n고민을 해결해요.',
-      '여러 사람들과 함께 프로젝트를 진행하며\n소프트 스킬을 개발하고, 효과적으로\n협업하는 방법을 배워요.',
-      '프로젝트 결과물은 기업, 대학원, 공모전\n등에 지원하거나 참여할 때 자신의 기술을\n보여주는 물리적 근거로 사용해요.',
-    ],
-  },
-  {
-    title: `Study & Networking`,
-    height: 210,
-    benefits: [
-      '전공, CS, 언어, 직무와 관련된\n스터디를 통해 전문성을 갖춰요.',
-      '정기 스터디를 통해 지속적으로 성장해요.',
-      '비슷한 관심사를 갖고 있는 사람과\n어울리며 인적 네트워크를 연결해요.',
-      '사람들과의 네트워크는 미래의 협업과\n취업에서 유리하게 작용해요.',
-      '졸업 프로젝트를 진행할 때 원활하게\n팀원을 조직해요.',
-    ],
-  },
-  {
-    title: 'Entertainment',
-    height: 200,
-    benefits: [`공부할 때 열심히 공부하고, \n놀 때 신나게 즐겨요.`, '월 1회 진행되는 \nLEETS DAY!', 'LEET’S GO PICNIC'],
-  },
-];
