--- utils/mobileViewports.tsx
@@ -1,4 +1,0 @@
-import facepaint from 'facepaint';
-
-const breakpoints = [360, 375, 390, 412];
-export default facepaint(breakpoints.map(breakpoint => `@media(min-width: ${breakpoint}px)`));
