--- utils/persistor.tsx
@@ -1,4 +1,0 @@
-import persistStore from 'redux-persist/es/persistStore';
-import store from '../store';
-
-export default persistStore(store);
