--- utils/timelineData.tsx
@@ -1,7 +1,0 @@
-export default {
-  지원기간: '23.03.02 ~ 23.03.10',
-  '면접 대상자 발표': '23.03.12',
-  면접: '23.03.13 ~ 23.03.14',
-  '최종 합격자 발표': '23.03.16',
-  '정규 활동 시작': '23.03.20',
-};
