--- utils/viewports.tsx
@@ -1,4 +1,0 @@
-import facepaint from 'facepaint';
-
-const breakpoints = [820, 1180, 1440];
-export default facepaint(breakpoints.map(breakpoint => `@media(min-width: ${breakpoint}px)`));
